// src/5_entities/Age/model/types/Age.ts
export enum Age {
  Underaged = "underaged",
  Age18 = "18",
  Age19 = "19",
  Age20 = "20",
  Age21 = "21",
  Age22 = "22",
  Age23 = "23",
  Age24 = "24",
  Age25 = "25",
  Age26 = "26",
  Age27 = "27",
  Age28 = "28",
  Age29 = "29",
  Age30 = "30",
  Age31 = "31",
  Age32 = "32",
  Age33 = "33",
  Age34 = "34",
  Age35 = "35",
  Age36 = "36",
  Age37 = "37",
  Age38 = "38",
  Age39 = "39",
  Age40 = "40",
  Age41 = "41",
  Age42 = "42",
  Age43 = "43",
  Age44 = "44",
  Age45 = "45",
  Age46 = "46",
  Age47 = "47",
  Age48 = "48",
  Age49 = "49",
  Age50 = "50",
  Age51 = "51",
  Age52 = "52",
  Age53 = "53",
  Age54 = "54",
  Age55 = "55",
  Age56 = "56",
  Age57 = "57",
  Age58 = "58",
  Age59 = "59",
  Age60 = "60",
  Elder = "elder",
}